requirements.txt
streamlit
streamlit-drawable-canvas
Pillow
python-dotenv
google.generativeai
numpy